pyproject.toml
[project]
name = "NeNeRobo"
version = "0.1.0"
description = ""
authors = [
    {name = "BlueGlassBlock", email = "blueglassblock@outlook.com"},
]
dependencies = [
    "kayaku~=0.5.0",
    "ichika[graia]<0.1.0,>=0.0.8",
    "githubkit[auth-oauth-device]~=0.10.0",
    "graiax-playwright>=0.2.4",
    "rich>=12.6.0",
    "mdit-py-emoji>=0.1.0",
    "markdown-it-py[linkify,plugins]>=2.1.0",
    "msgspec>=0.9.0",
    "graiax-shortcut>=0.3.0",
    "sphobjinv>=2.2.2",
    "aiosqlite~=0.19.0",
]
requires-python = ">=3.10"
license = {text = "AGPL-3.0"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.8.0",
    "isort>=5.10.1",
]

[tool.isort]
profile = "black"